tried to add text

+++ common/text2D.hpp
@@ -2,7 +2,7 @@
 #define TEXT2D_HPP
 
 void initText2D(const char * texturePath);
-void printText2D(const char * text, int x, int y, int size);
+void printText2D(const char * text, float x, float y, float size);
 void cleanupText2D();
 
 #endif
